Refactor and update molecule processing and preparation strategies
- Cleaned up whitespace and formatting in MoleculeProcessing.py and Preparation.py for better readability.
- Updated error handling in various functions to remove type ignores and ensure proper error messages.
- Modified the PDB file cleaning process to ensure consistent formatting and added necessary headers.
- Enhanced the validation checks in Validation.py to ensure proper parsing of molecule files.
- Adjusted test cases in test files to reflect changes in expected values and function signatures.
- Updated mypy configuration to remove unnecessary warnings and streamline type checking.
- Changed references in the DNN optimizer configuration to align with updated target naming conventions.

+++ OCDocker/OCScore/Dimensionality/future/utils.py
@@ -184,7 +184,7 @@ def ramp_weight(
 
     if ramp_type == "sigmoid":
         # Sigmoid ramp grows slowly then saturates.
-        return float(target) * (1.0 / (1.0 + np.exp(-10.0 * (progress - 0.5))))
+        return float(target) * float(1.0 / (1.0 + np.exp(-10.0 * (progress - 0.5))))
 
     return float(target) * progress
